Only board author may update or delete their boards

+++ backend/Boards/api/v1/views.py
@@ -15,11 +15,16 @@ class BoardList(APIView):
     permission_classes = (IsAuthenticated,)
 
     def get(self, request, format=None):
-        boards = Board.objects.all()
+        # print(f"{request.user = }")
+        # print(f"{request.user.id = }")
+        # get current user boards
+        boards = Board.objects.filter(created_by=request.user.id)
         serialized_boards = BoardSerializer(instance=boards, many=True)
         return Response(data=serialized_boards.data, status=status.HTTP_200_OK)
 
     def post(self, request, format=None):
+        # use user id from request data as required board creator id
+        request.data["created_by"] = request.user.id
         serialized_board = BoardSerializer(data=request.data)
         if serialized_board.is_valid():
             serialized_board.save()
@@ -47,21 +52,27 @@ class BoardDetails(APIView):
 
     def patch(self, request, pk, format=None):
         board = self.get_object(pk)
-        serialized_board = BoardSerializer(board, data=request.data, partial=True)
-        if serialized_board.is_valid():
-            serialized_board.save()
-            return Response(serialized_board.data)
-        return Response(serialized_board.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.email == str(board.created_by):
+            serialized_board = BoardSerializer(board, data=request.data, partial=True)
+            if serialized_board.is_valid():
+                serialized_board.save()
+                return Response(serialized_board.data)
+            return Response(serialized_board.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
     def put(self, request, pk, format=None):
         board = self.get_object(pk)
-        serialized_board = BoardSerializer(board, data=request.data)
-        if serialized_board.is_valid():
-            serialized_board.save()
-            return Response(serialized_board.data)
-        return Response(serialized_board.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.email == str(board.created_by):
+            serialized_board = BoardSerializer(board, data=request.data)
+            if serialized_board.is_valid():
+                serialized_board.save()
+                return Response(serialized_board.data)
+            return Response(serialized_board.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
 
     def delete(self, request, pk, format=None):
         board = self.get_object(pk)
-        board.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.email == str(board.created_by):
+            board.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response(status=status.HTTP_401_UNAUTHORIZED)
